requirements.txt
fastapi
uvicorn
azure-cognitiveservices-speech
azure-ai-textanalytics
python-dotenv
